ComputeAsync returns an async enumerable. Fix namespace typo in EnumerableExtensions.
There's a perf cost to using the async enumerable instead of the
individual properties for each part, but it allows me to measure the
computation cost for each part separately where applicable, without
the implementation awkwardness of distinct part 1/2 computation
methods when the ideal path overlaps. Mainly helpful for perf tuning
when part 2 is computationally out of balance with part 1.