2023 Day 8 perf
If I'm going to make assumptions about the shape of the input data, I
might as well go all the way. Assuming periodicity along multiples of
the length of the instruction set allows for drastically faster calculation.

I also swapped the instructions from characters to array indexes, to
avoid a lot of equality checking in switch statements.