2023 Day 16 perf
Precalculating all of the splitters, since a splitter energizes the same tiles
regardless of how it's entered. That way my "runtime" only needs me to
trace as far as the first splitter. I'm still doing a lot of redundant work in
this precalculation, but it's 4-5x faster than my previous entry, so yay!