Fix a bug in IEnumerable<T>.Cluster()
When the outer enumerable is evaluated before the inner enumerables,
the inner enumerables never allocate their set of lines. So I'm forcing inner
enumerable allocation. I'm not sure this is the best way to handle this,
since it might force enumerable evaluation earlier than could be ideal in
some situations, but it should be fine for my intended usage.