2023 Day 21
Once again, partly a coding problem, but moreso a data analysis problem.
The solution for part 2 depends on the design of the input data, in order
for it to be calculated in a reasonable amount of time.

I know I can speed up the evaluation of part 2 with more assumptions,
since the calculation is already rife with assumptions. The only trick is
figuring out how many plots are unreachable, and then counting even or
odd step for inner or outer diamond plots. And math. Maybe later.