2023 Day 24, part two
It's not pretty, but it's done. I jumped through a lot of hoops trying to
avoid uneven BigInteger division, and it's quite possible that this solution
won't work for all Advent of Code puzzle data... but it works for mine, and
I can worry about creating a generic Gaussian reducer that handles large
values another day.